Make e2e tests pass
Change naming and made lots of small tweaks so that all of the e2e tests that we expect to pass, pass. We also commented out/removed tests that we don't expect to pass, due to no implementation of mongo. We additionally removed the reading from file functionality in exchange for dynamically generating ASCII strings of arbitrary length.
Co-authored-by: jpwalbran <walbr042@morris.umn.edu>